requirements.txt
pillow==12.1.0
CairoSVG==2.8.2
PySide6==6.10.2
pyinstaller==6.18.0